requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-telegram-bot==20.8
celery==5.3.6
redis==5.0.4
Flask-Admin==1.6.1
